activity with button

+++ app/src/main/java/ru/touchin/spizdev/models/WiFiScan.kt
@@ -3,6 +3,6 @@ package ru.touchin.spizdev.models
 data class WiFiScan(
     val name: String,
     val macAddress: String,
-    val frequencyInMhz: Double,
-    val levelInDb: Double
+    val frequencyInMhz: Int,
+    val levelInDb: Int
 )
